==> 2_muestra5000.ipynb <==
import pandas as pd
scrape = pd.read_csv("tweets_scrape.csv")

scrape.head()

# escribe tweet_text de fila 0
print(scrape.loc[0, "tweet_text"])

def extraer_fecha(texto):
    patron = r"\d{1,2}:\d{2} [AP]M · [A-Za-z]{3,9} \d{1,2}, \d{4}"
    match = re.search(patron, texto)
    return match.group(0) if match else None

# Aplicar la función
tweets_df["fecha_raw"] = tweets_df["tweet_text"].apply(extraer_fecha)
tweets_df["fecha_datetime"] = pd.to_datetime(tweets_df["fecha_raw"], errors='coerce')

# 4. Análisis temporal
print("🔍 Intervalo temporal:")
print("Fecha mínima:", tweets_df["fecha_datetime"].min())
print("Fecha máxima:", tweets_df["fecha_datetime"].max())


# Montar Google Drive
from google.colab import drive
drive.mount('/content/drive')

# Librerías necesarias
import pandas as pd
import re

# 1. Cargar archivo de notas en español
notes_spanish_path = '/content/drive/MyDrive/TFM/notesdataspanish_.csv'
notes = pd.read_csv(notes_spanish_path)

# Eliminar notas sin tweetId y eliminar duplicados por tweetId
notes = notes.dropna(subset=["tweetId"])
notes = notes.drop_duplicates(subset=["tweetId"])

# Seleccionar muestra aleatoria de 5000 tweetIds
muestra_ids = notes["tweetId"].sample(n=5000, random_state=42).astype(str).tolist()



import pandas as pd

# Crear CSV con la muestra
muestra_df = pd.DataFrame({"tweet_id": muestra_ids})
muestra_df.to_csv("muestra_5000.csv", index=False)


print(muestra_ids[:10])  # Muestra los primeros 10 tweetId de la muestra


# 2. Cargar archivo de tweets scrapeado
tweets_scrape_path = '/content/drive/MyDrive/TFM/                      '
tweets_df = pd.read_csv(tweets_scrape_path)

# Limpiar duplicados por tweet_id
tweets_df = tweets_df.drop_duplicates(subset=["tweet_id"])

# Filtrar para quedarnos solo con los tweet_id seleccionados
tweets_df = tweets_df[tweets_df["tweet_id"].astype(str).isin(muestra_ids)]

# Guardar archivo limpio
tweets_df.to_csv("tweets_scrape_limpio.csv", index=False)

# 3. Extraer fecha desde tweet_text
def extraer_fecha(texto):
    patron = r"\d{1,2}:\d{2} [AP]M · [A-Za-z]{3,9} \d{1,2}, \d{4}"
    match = re.search(patron, texto)
    return match.group(0) if match else None

# Aplicar la función
tweets_df["fecha_raw"] = tweets_df["tweet_text"].apply(extraer_fecha)
tweets_df["fecha_datetime"] = pd.to_datetime(tweets_df["fecha_raw"], errors='coerce')

# 4. Análisis temporal
print("🔍 Intervalo temporal:")
print("Fecha mínima:", tweets_df["fecha_datetime"].min())
print("Fecha máxima:", tweets_df["fecha_datetime"].max())

# Distribución por mes
tweets_df["mes"] = tweets_df["fecha_datetime"].dt.to_period("M")
conteo_mensual = tweets_df["mes"].value_counts().sort_index()

print("\n📈 Tweets por mes:")
print(conteo_mensual)


# Montar Google Drive
from google.colab import drive
drive.mount('/content/drive')

# Librerías necesarias
import pandas as pd
import re

# 1. Cargar archivo de notas en español
twe_spanish_path = '/content/drive/MyDrive/TFM/notesdataspanish_.csv'
twe = pd.read_csv(twe_spanish_path)
